19330 【模具排产-算法】没有生成未排产结果表

+++ APS-Modules/aps-factory-engine/src/main/java/com/zlt/aps/factory/utils/NoProductionPlanUtils.java
@@ -28,7 +28,7 @@ public class NoProductionPlanUtils {
         noProductionPlan.setReason(unProductionReason);
         Integer needProductionQty = productionPlan.getFactProdReqQty();
         Integer plannedQty = sumProductionMap.getOrDefault(monthPlanId,0);
-        long unProductionQty = needProductionQty - plannedQty;
+        int unProductionQty = needProductionQty - plannedQty;
         noProductionPlan.setUnProductionQty(unProductionQty);
         if (!needProductionQty.equals(plannedQty)) {
           list.add(noProductionPlan);
